ch4_sot_and_templates/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.2
PyYAML==6.0.2
requests==2.32.4
urllib3==2.4.0
